reorganize flask app
