+++ cg/server/auto.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+from .app import create_app
+
+
+app = create_app(config='cg.server.config')
